requirements.txt
instagrapi
python-dotenv
discord.py
schedule
